fix WebTransportStream test
The code had an accidental cast from long to uint.

+++ src/libraries/System.Net.Http/tests/FunctionalTests/WebTransport/WebTransportStreamTests.cs
@@ -59,7 +59,7 @@ public sealed class WebTransportStreamTests : WebTransportTestBase
         var theoryData = new TheoryData<WebTransportStreamType, long>();
         foreach (WebTransportStreamType streamType in Enum.GetValues(typeof(WebTransportStreamType)))
         {
-            foreach (uint errorCode in s_invalidErrorCodes)
+            foreach (long errorCode in s_invalidErrorCodes)
             {
                 theoryData.Add(streamType, errorCode);
             }
